xixi06261128

+++ communityiOS/CommodityInfo.h
@@ -15,7 +15,7 @@
 @property (strong,nonatomic) NSString *commodity_id;
 @property (strong,nonatomic) NSString *comm_name;
 @property (strong,nonatomic) NSString *comm_unit; //商品单位
-@property (assign,nonatomic) float comm_price;
+@property (assign,nonatomic) double comm_price;
 @property (strong,nonatomic) NSString *comm_state; //商品状态：上架、下架、删除
 @property (strong,nonatomic) NSString *comm_regdate; //商品登记时间
 @property (strong,nonatomic) NSString *comm_desc;  //商品描述
